karma: restructure tests to enable auto rerun
