+++ packages/base/package.json
@@ -18,17 +18,16 @@
   "scripts": {
     "clean:src": "rimraf lib",
     "clean": "npm run clean:src",
-    "build:test": "tsc --project test/src && webpack --config test/webpack.conf.js",
     "build:src": "tsc --project src",
     "build": "npm run build:src",
     "lint": "tslint --project tslint.json --format stylish",
     "test": "npm run test:unit",
-    "test:coverage": "npm run build:test && webpack --config test/webpack-cov.conf.js && karma start test/karma-cov.conf.js",
     "test:unit": "npm run test:unit:firefox && npm run test:unit:chrome",
-    "test:unit:default": "npm run build:test && karma start test/karma.conf.js --log-level debug",
+    "test:unit:default": "karma start --log-level debug",
     "test:unit:firefox": "npm run test:unit:default -- --browsers=Firefox",
     "test:unit:chrome": "npm run test:unit:default -- --browsers=Chrome",
     "test:unit:ie": "npm run test:unit:default -- --browsers=IE",
+    "test:unit:watch": "npm run test:unit:default -- --browsers=Chrome --singleRun=false",
     "prepublish": "npm run clean && npm run build",
     "watch": "tsc -w --project src"
   },
@@ -39,13 +38,14 @@
     "chai-as-promised": "^7.0.0",
     "expect.js": "^0.3.1",
     "istanbul-instrumenter-loader": "^2.0.0",
-    "karma": "^1.6.0",
+    "karma": "^3.1.3",
     "karma-chrome-launcher": "^2.0.0",
     "karma-coverage": "^1.0.0",
     "karma-firefox-launcher": "^1.0.1",
     "karma-ie-launcher": "^1.0.0",
     "karma-mocha": "^1.3.0",
     "karma-mocha-reporter": "^2.2.3",
+    "karma-typescript": "^3.0.13",
     "karma-webpack": "^2.0.3",
     "mocha": "^3.3.0",
     "npm-run-all": "^1.5.1",
