--- packages/base/test/karma.conf.js
@@ -1,13 +1,0 @@
-module.exports = function (config) {
-  config.set({
-    basePath: '..',
-    frameworks: ['mocha'],
-    reporters: ['mocha'],
-    files: ['test/build/bundle.js'],
-    port: 9876,
-    colors: true,
-    singleRun: true,
-    logLevel: config.LOG_INFO,
-    browserNoActivityTimeout: 30000
-  });
-};
